backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
pytest==7.4.3
Flask-RESTful==0.3.10
Werkzeug==3.0.1
email-validator==2.1.0.post1
